backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
django-filter>=23.5
drf-spectacular>=0.27.0
Faker>=22.0.0
openai>=1.0.0
